+++ doc/sphinx/changelog.rst
@@ -6,7 +6,7 @@ This page documents what's new in each updated version of TurboCtl.
 Version 1.1.1
 -------------
 
-* Moved TurboCtl documentation to `Read the Docs <https://turboctl.readthedocs.io/en/latest/index.html>`_.
+* Moved TurboCtl documentation to `Read the Docs`_.
   Previously documentation could be viewed online via https://html-preview.github.io/, but this broke some links.
 * Added the :option:`-d` command line argument.
 
@@ -18,5 +18,3 @@ Version 1.1.0
 * Made TurboCtl available on PyPI_ and updated the installation instructions.
 * Added the :option:`docs` and :option:`reset` commands.
 * Made several other small improvements and fixes.
-
-.. _PyPI: https://pypi.org/project/turboctl/